Separado en componentes el movie
